fix: Address possible crash with overlapping sidebar entries (#56)
`SidebarItem` struct identifiers were not guaranteed to be unique across
the app as they only used the folder URL and not the owner URL. This
change uses the new `Details.Identifier` as an identifier for sidebar
items as it's guaranteed to be unique within the app. It also tidies up
some of the property names for greater consistency between `SidebarItem`
and `Details`.