feat: Show folder links in the toolbar (#159)
Previously links were only shown at the bottom of the grid view if links
had been defined. This change moves them to an ever-present drop-down in
the toolbar. This should hopefully improve discoverability of the
feature and provide a natural place for users to edit links.

This turned out to be a little more nuanced than I had first expected as
I ended up reworking the `StorePublisher` (now `StoreFilePublisher`) to
load the initial state and updated the model to rely almost exclusively
on this to load settings on change. It does a little more work than
strictly necessary as, in order to avoid caching too much in the Combine
stream, it reloads all relevant settings files every time any one
changes. We can revisit this if it proves an over-simplification.

While it might seem counter intuitive that `SelectionModel` remains an
`ObservableObject` and doesn't adopt `Observable`, I believe the
`.focusedSceneObject` view modifier still doesn't work correctly with
`Observable`, so I'm sticking with the old approach just in case.